+++ nodekit-server/nodekit_server/routers/tags.py
@@ -1,17 +1,16 @@
 """Tag routes for nodekit-server."""
 
 from typing import Annotated
-from uuid import UUID, uuid4
+from uuid import uuid4
 
 import fastapi
 import sqlmodel
-from sqlalchemy import and_, or_
 
 import nodekit.server.contracts as contracts
-from nodekit.server.pagination import decode_timestamp_id_cursor, encode_timestamp_id_cursor
 from nodekit.server.values import UserId
 from nodekit_server.auth import UserDep
 from nodekit_server.deps import SessionDep
+from nodekit_server.pagination import apply_timestamp_id_page_cursor, page_records
 from nodekit_server.records import TagRecord, as_utc
 
 
@@ -76,31 +75,6 @@ def _dedupe_names(names: list[str] | None) -> tuple[str, ...]:
     return tuple(deduped)
 
 
-def _apply_page_cursor(
-    statement,
-    page_cursor: str,
-):
-    try:
-        cursor = decode_timestamp_id_cursor(page_cursor)
-        cursor_tag_id = UUID(cursor.id)
-    except ValueError as exc:
-        raise fastapi.HTTPException(
-            status_code=fastapi.status.HTTP_400_BAD_REQUEST,
-            detail="Invalid page cursor.",
-        ) from exc
-
-    cursor_timestamp = cursor.timestamp_created.replace(tzinfo=None)
-    return statement.where(
-        or_(
-            sqlmodel.col(TagRecord.timestamp_created) < cursor_timestamp,
-            and_(
-                sqlmodel.col(TagRecord.timestamp_created) == cursor_timestamp,
-                sqlmodel.col(TagRecord.tag_id) < cursor_tag_id,
-            ),
-        )
-    )
-
-
 def _create_tag_record(
     session: sqlmodel.Session,
     user_id: UserId,
@@ -161,7 +135,12 @@ def list_tags(
         statement = statement.where(sqlmodel.col(TagRecord.is_archived) == False)  # noqa: E712
 
     if query.page_cursor is not None:
-        statement = _apply_page_cursor(statement=statement, page_cursor=query.page_cursor)
+        statement = apply_timestamp_id_page_cursor(
+            statement=statement,
+            page_cursor=query.page_cursor,
+            timestamp_column=TagRecord.timestamp_created,
+            id_column=TagRecord.tag_id,
+        )
 
     statement = statement.order_by(
         sqlmodel.col(TagRecord.timestamp_created).desc(),
@@ -169,18 +148,15 @@ def list_tags(
     )
     statement = statement.limit(query.max_items + 1)
     tag_records = session.exec(statement).all()
-    page_records = tag_records[: query.max_items]
-
-    next_page_cursor = None
-    if len(tag_records) > query.max_items and page_records:
-        last_record = page_records[-1]
-        next_page_cursor = encode_timestamp_id_cursor(
-            timestamp_created=as_utc(last_record.timestamp_created),
-            id=last_record.tag_id,
-        )
+    records_page, next_page_cursor = page_records(
+        records=list(tag_records),
+        max_items=query.max_items,
+        timestamp_attr="timestamp_created",
+        id_attr="tag_id",
+    )
 
     return contracts.ListTagsResponse(
-        items=[_tag_item(tag_record) for tag_record in page_records],
+        items=[_tag_item(tag_record) for tag_record in records_page],
         next_page_cursor=next_page_cursor,
     )
