+++ tests/conftest.py
@@ -1,10 +1,62 @@
+import importlib
+import os
+import sys
+from collections.abc import Iterator
 from pathlib import Path
+from types import ModuleType
 
 import pytest
+from fastapi.testclient import TestClient
+from sqlmodel import SQLModel
 
 import nodekit as nk
 
 ASSET_DIR = Path(__file__).parent / "assets"
+SERVER_ROOT = Path(__file__).parents[1] / "nodekit-server"
+
+
+@pytest.fixture(scope="session")
+def server_main(tmp_path_factory: pytest.TempPathFactory) -> Iterator[ModuleType]:
+    tmp_path = tmp_path_factory.mktemp("nodekit-server")
+    env_keys = [
+        "NODEKIT_SERVER_DATABASE_URL",
+        "NODEKIT_SERVER_ASSET_STORE_DIR",
+        "NODEKIT_SERVER_BOOTSTRAP_ADMIN_API_TOKEN",
+    ]
+    previous_env = {key: os.environ.get(key) for key in env_keys}
+    os.environ["NODEKIT_SERVER_DATABASE_URL"] = f"sqlite:///{tmp_path / 'server.db'}"
+    os.environ["NODEKIT_SERVER_ASSET_STORE_DIR"] = str(tmp_path / "asset-store")
+    os.environ["NODEKIT_SERVER_BOOTSTRAP_ADMIN_API_TOKEN"] = "test-token"
+    sys.path.insert(0, str(SERVER_ROOT))
+
+    for module_name in list(sys.modules):
+        if module_name == "nodekit_server" or module_name.startswith("nodekit_server."):
+            del sys.modules[module_name]
+    SQLModel.metadata.clear()
+
+    module = importlib.import_module("nodekit_server.main")
+    yield module
+
+    for module_name in list(sys.modules):
+        if module_name == "nodekit_server" or module_name.startswith("nodekit_server."):
+            del sys.modules[module_name]
+    SQLModel.metadata.clear()
+    try:
+        sys.path.remove(str(SERVER_ROOT))
+    except ValueError:
+        pass
+    for key, value in previous_env.items():
+        if value is None:
+            os.environ.pop(key, None)
+        else:
+            os.environ[key] = value
+
+
+@pytest.fixture
+def authenticated_client(server_main: ModuleType) -> Iterator[TestClient]:
+    with TestClient(server_main.app) as client:
+        client.headers.update({"Authorization": "Bearer test-token"})
+        yield client
 
 
 def _make_graph_with_assets() -> nk.Graph:
